+++ crates/node/src/main.rs
@@ -2,18 +2,16 @@
 #![warn(missing_docs)]
 #![warn(clippy::unwrap_used)]
 
+mod cli;
+mod service;
+mod util;
+
 use std::sync::Arc;
 
 use anyhow::Context;
 use clap::Parser;
 use mc_block_import::BlockImporter;
-mod cli;
-mod service;
-mod util;
 
-use crate::cli::NetworkType;
-use crate::service::L1SyncService;
-use cli::RunCmd;
 use mc_db::DatabaseService;
 use mc_mempool::{GasPriceProvider, L1DataProvider, Mempool};
 use mc_metrics::MetricsService;
@@ -21,8 +19,12 @@ use mc_rpc::providers::{AddTransactionProvider, ForwardToProvider, MempoolAddTxP
 use mc_telemetry::{SysInfo, TelemetryService};
 use mp_convert::ToFelt;
 use mp_utils::service::{Service, ServiceGroup};
-use service::{BlockProductionService, RpcService, SyncService};
+
 use starknet_providers::SequencerGatewayProvider;
+
+use cli::{NetworkType, RunCmd};
+use service::L1SyncService;
+use service::{BlockProductionService, RpcService, SyncService};
 
 const GREET_IMPL_NAME: &str = "Madara";
 const GREET_SUPPORT_URL: &str = "https://github.com/madara-alliance/madara/issues";
@@ -34,7 +36,8 @@ async fn main() -> anyhow::Result<()> {
     crate::util::raise_fdlimit();
 
     let mut run_cmd: RunCmd = RunCmd::parse();
-    let chain_config = run_cmd.network.chain_config();
+
+    let chain_config = run_cmd.get_config()?;
 
     let node_name = run_cmd.node_name_or_provide().await.to_string();
     let node_version = env!("DEOXYS_BUILD_VERSION");